AEA-293: Refactored handler code to remove duplication.
