--- coordinator/src/services/request-body-parser.ts
@@ -1,5 +1,0 @@
-import Hapi from "@hapi/hapi";
-
-export function parse(request: Hapi.Request): unknown {
-    return request.payload || {}
-}
